requirements.txt
ttkbootstrap>=1.10.1
requests>=2.31.0
Pillow>=10.0.0
